RDoc can be generated from any GLI-based command
Every GLI-based command will now have a secret "rdoc" option
that will output command_name.rdoc in the current directory;
this contains the rdoc description of the command's functioning.

I realize this isn't ideal from an architecture point of view,
but this is still useful enough to go ahead and put in now.

Fixes #4