removed default use of OpenStruct/Option
I need to treat the options as a Hash, and OpenStruct doesn't
allow that.  Deferred the conversion until the end and made it optional.
I think it will be more surising to people trying to manipulate it as a Hash
and failing than to those who use it as an OpenStruct. 

Added an option to use it for those who want it.

Also fixing some horrible bugs I didn't think could happen in Ruby, namely
that it seems you must initialize class variables.