Fixed bugs with config
Closes #43
Closes #10

This design is starting to crack some, but this was the cleanest way to do it.
Basically, look into the configuration and set the default values of
flags/switches *before* we parse the command line.  This simplifies
things in the parsing a bit, but the flag/switch hierarchy is a bit messed up
now.