Clarify CLI syntax for multi-value arguments
GLI will not strip comma characters from the end of arguments. If a user
is following the help syntax for a multi-value argument, the first to
N-1 values parsed will be incorrect (have an unstripped comma at the
end of the value). Solve this by printing the argument name only once,
followed by '...' directly. This is similar to the behaviour of git.

Signed-off-by: Florian Pritz <bluewind@xinu.at>